# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API Barantin",
        "description": "Data Request",
        "contact": {
            "email": "nandaputraservice@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Api Admin Barantin"
        },
        {
            "url": "http://localhost:8000/api/v2",
            "description": "Api User Barantin Test"
        }
    ],
    "paths": {
...
        "/ppjk/{take}": {
            "get": {
                "tags": [
                    "PPJK Admin"
                ],
                "summary": "Get PPJK",
                "description": "Get PPJK",
                "operationId": "getPpjkAdmin",
                "parameters": [
                    {
                        "name": "take",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Data Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/ppjk/{ppjk_id}/detil": {
            "get": {
                "tags": [
                    "PPJK Admin"
                ],
                "summary": "Get Detail PPJK",
                "description": "Get Detail PPJK",
                "operationId": "getDetailPpjkAdmin",
                "parameters": [
                    {
                        "name": "ppjk_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Data Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/ppjk/cek-npwp": {
            "get": {
                "tags": [
                    "PPJK Admin"
                ],
                "summary": "Cek NPWP PPJK",
                "description": "Cek NPWP PPJK",
                "operationId": "cekNpwpPpjk",
                "parameters": [
                    {
                        "name": "kd_upt",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "npwp_ppjk",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "npwp_pj",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Data Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/user/cabang": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get Cabang Perusahaan Induk",
                "description": "Get Cabang Perusahaan Induk",
                "operationId": "getCabangPerusahaanInduk",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Data Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/user/cabang/{barantin_cabang_id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get Detail Cabang Perusahaan Induk",
                "description": "Get Detail Cabang Perusahaan Induk",
                "operationId": "getDetailCabangPerusahaanInduk",
                "parameters": [
                    {
                        "name": "baratin_cabang_id",
                        "in": "path",
                        "description": "ID of the Barantin Cabang",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Data Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
        "/user/ppjk": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get PPJK",
                "description": "Get PPJK Pengguna Jasa",
                "operationId": "getPpjk",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Data Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/user/ppjk/{ppjk_id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get Detail PPJK Pengguna Jasa",
                "description": "Get Detail PPJK",
                "operationId": "getDetailPpjk",
                "parameters": [
                    {
                        "name": "ppjk_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Data Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Tambahkan Kata \"Bearer\" Sebelum token contoh: (Bearer xx|497zSWISvECINqxxxxxxxxxxxxxxx)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}